analysis-core/models/server/requirements.txt
essentia==2.1b6.dev1034
grpcio==1.60.0
grpcio-tools==1.60.0
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
faiss-cpu==1.7.4
protobuf==4.25.1
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1